use NUInteger instead of NSInteger for all possible scenarios

+++ DTModelStorage/Memory/DTMemoryStorage.m
@@ -28,25 +28,25 @@
 #import "DTStorageUpdate.h"
 #import "DTSectionModel.h"
 
-@interface DTMemoryStorage()
+@interface DTMemoryStorage ()
 @property (nonatomic, strong) DTStorageUpdate * currentUpdate;
 @property (nonatomic, retain) NSMutableDictionary * searchingBlocks;
 @end
 
 @implementation DTMemoryStorage
 
-+(instancetype)storage
++ (instancetype)storage
 {
     DTMemoryStorage * storage = [self new];
-    
+
     storage.sections = [NSMutableArray array];
-    
+
     storage.loggingEnabled = YES;
-    
+
     return storage;
 }
 
--(NSMutableDictionary *)searchingBlocks
+- (NSMutableDictionary *)searchingBlocks
 {
     if (!_searchingBlocks)
     {
@@ -55,38 +55,40 @@
     return _searchingBlocks;
 }
 
--(id)objectAtIndexPath:(NSIndexPath *)indexPath
+- (id)objectAtIndexPath:(NSIndexPath *)indexPath
 {
     id <DTSection> sectionModel = nil;
     if (indexPath.section >= self.sections.count)
     {
         return nil;
     }
-    else {
+    else
+    {
         sectionModel = [self sections][indexPath.section];
-        if (indexPath.item>= [sectionModel numberOfObjects])
+        if (indexPath.item >= [sectionModel numberOfObjects])
         {
             return nil;
         }
     }
-    
+
     return [sectionModel.objects objectAtIndex:indexPath.row];
 }
 
--(id)supplementaryModelOfKind:(NSString *)kind forSectionIndex:(NSInteger)sectionNumber
+- (id)supplementaryModelOfKind:(NSString *)kind forSectionIndex:(NSUInteger)sectionNumber
 {
     DTSectionModel * sectionModel = nil;
     if (sectionNumber >= self.sections.count)
     {
         return nil;
     }
-    else {
+    else
+    {
         sectionModel = [self sections][sectionNumber];
     }
     return [sectionModel supplementaryModelOfKind:kind];
 }
 
--(void)setSupplementaries:(NSArray *)supplementaryModels forKind:(NSString *)kind
+- (void)setSupplementaries:(NSArray *)supplementaryModels forKind:(NSString *)kind
 {
     [self startUpdate];
     if (!supplementaryModels || [supplementaryModels count] == 0)
@@ -97,9 +99,9 @@
         }
         return;
     }
-    [self getValidSection:([supplementaryModels count]-1)];
-    
-    for (int sectionNumber = 0; sectionNumber < [supplementaryModels count]; sectionNumber++)
+    [self getValidSection:([supplementaryModels count] - 1)];
+
+    for (NSUInteger sectionNumber = 0; sectionNumber < [supplementaryModels count]; sectionNumber++)
     {
         DTSectionModel * section = self.sections[sectionNumber];
         [section setSupplementaryModel:supplementaryModels[sectionNumber] forKind:kind];
@@ -109,22 +111,22 @@
 
 #pragma mark - search
 
--(void)setSearchingBlock:(DTModelSearchingBlock)searchingBlock
-           forModelClass:(Class)modelClass
+- (void)setSearchingBlock:(DTModelSearchingBlock)searchingBlock
+            forModelClass:(Class)modelClass
 {
     NSParameterAssert(searchingBlock);
     NSParameterAssert(modelClass);
-    
+
     self.searchingBlocks[[self classStringForClass:modelClass]] = searchingBlock;
 }
 
--(instancetype)searchingStorageForSearchString:(NSString *)searchString
-                                 inSearchScope:(NSInteger)searchScope
+- (instancetype)searchingStorageForSearchString:(NSString *)searchString
+                                  inSearchScope:(NSUInteger)searchScope
 {
     DTMemoryStorage * storage = [[self class] storage
-                                          ];
-    
-    for (int sectionNumber = 0; sectionNumber < [self.sections count]; sectionNumber++)
+    ];
+
+    for (NSUInteger sectionNumber = 0; sectionNumber < [self.sections count]; sectionNumber++)
     {
         DTSectionModel * searchSection = [self filterSection:self.sections[sectionNumber]
                                             withSearchString:searchString
@@ -134,32 +136,33 @@
             [storage.sections addObject:searchSection];
         }
     }
-    
+
     return storage;
 }
 
--(DTSectionModel *)filterSection:(DTSectionModel*)section
-                withSearchString:(NSString *)searchString
-                     searchScope:(NSInteger)searchScope
+- (DTSectionModel *)filterSection:(DTSectionModel *)section
+                 withSearchString:(NSString *)searchString
+                      searchScope:(NSInteger)searchScope
 {
     NSMutableArray * searchResults = [NSMutableArray array];
-    for (int row = 0; row < section.objects.count ; row++)
+    for (NSUInteger row = 0; row < section.objects.count; row++)
     {
 #pragma clang diagnostic push
 #pragma clang diagnostic ignored "-Wdeprecated-declarations"
 
         NSObject <DTModelSearching> * item = section.objects[row];
-        
+
         if (self.searchingBlocks[[self classStringForClass:item.class]])
         {
             DTModelSearchingBlock block = self.searchingBlocks[[self classStringForClass:item.class]];
-            
-            if (block && block(item,searchString,searchScope,section))
+
+            if (block && block(item, searchString, searchScope, section))
             {
                 [searchResults addObject:item];
             }
         }
-        else { // Check for deprecated DTModelSearching protocol
+        else
+        { // Check for deprecated DTModelSearching protocol
             if ([item respondsToSelector:@selector(shouldShowInSearchResultsForSearchString:inScopeIndex:)])
             {
                 BOOL shouldShow = [item shouldShowInSearchResultsForSearchString:searchString
@@ -184,12 +187,12 @@
 
 #pragma mark - Updates
 
--(void)startUpdate
+- (void)startUpdate
 {
     self.currentUpdate = [DTStorageUpdate new];
 }
 
--(void)finishUpdate
+- (void)finishUpdate
 {
     if ([self.delegate respondsToSelector:@selector(storageDidPerformUpdate:)])
     {
@@ -200,35 +203,35 @@
 
 #pragma mark - Adding items
 
--(void)addItem:(NSObject *)item
+- (void)addItem:(NSObject *)item
 {
     [self addItem:item toSection:0];
 }
 
--(void)addItem:(NSObject *)item toSection:(NSInteger)sectionNumber
+- (void)addItem:(NSObject *)item toSection:(NSUInteger)sectionNumber
 {
     [self startUpdate];
-    
+
     DTSectionModel * section = [self getValidSection:sectionNumber];
     NSUInteger numberOfItems = [section numberOfObjects];
     [section.objects addObject:item];
     [self.currentUpdate.insertedRowIndexPaths addObject:[NSIndexPath indexPathForRow:numberOfItems
                                                                            inSection:sectionNumber]];
-    
+
     [self finishUpdate];
 }
 
--(void)addItems:(NSArray *)items
+- (void)addItems:(NSArray *)items
 {
     [self addItems:items toSection:0];
 }
 
--(void)addItems:(NSArray *)items toSection:(NSInteger)sectionNumber
+- (void)addItems:(NSArray *)items toSection:(NSUInteger)sectionNumber
 {
     [self startUpdate];
-    
+
     DTSectionModel * section = [self getValidSection:sectionNumber];
-    
+
     for (id item in items)
     {
         NSUInteger numberOfItems = [section numberOfObjects];
@@ -236,44 +239,45 @@
         [self.currentUpdate.insertedRowIndexPaths addObject:[NSIndexPath indexPathForRow:numberOfItems
                                                                                inSection:sectionNumber]];
     }
-    
+
     [self finishUpdate];
 }
 
--(void)insertItem:(NSObject *)item toIndexPath:(NSIndexPath *)indexPath
+- (void)insertItem:(NSObject *)item toIndexPath:(NSIndexPath *)indexPath
 {
     [self startUpdate];
     // Update datasource
     DTSectionModel * section = [self getValidSection:indexPath.section];
-    
+
     if ([section.objects count] < indexPath.row)
     {
-        if (self.loggingEnabled) {
+        if (self.loggingEnabled)
+        {
             NSLog(@"DTMemoryStorage: failed to insert item for section: %ld, row: %ld, only %lu items in section",
-                  (long)indexPath.section,
-                  (long)indexPath.row,
-                  (unsigned long)[section.objects count]);
+                    (long)indexPath.section,
+                    (long)indexPath.row,
+                    (unsigned long)[section.objects count]);
         }
         return;
     }
     [section.objects insertObject:item atIndex:indexPath.row];
-    
+
     [self.currentUpdate.insertedRowIndexPaths addObject:indexPath];
-    
+
     [self finishUpdate];
 }
 
--(void)reloadItem:(NSObject *)item
+- (void)reloadItem:(NSObject *)item
 {
     [self startUpdate];
-    
+
     NSIndexPath * indexPathToReload = [self indexPathForItem:item];
-    
+
     if (indexPathToReload)
     {
         [self.currentUpdate.updatedRowIndexPaths addObject:indexPathToReload];
     }
-    
+
     [self finishUpdate];
 }
 
@@ -281,24 +285,26 @@
            withItem:(NSObject *)replacingItem
 {
     [self startUpdate];
-    
+
     NSIndexPath * originalIndexPath = [self indexPathForItem:itemToReplace];
     if (originalIndexPath && replacingItem)
     {
-        DTSectionModel *section = [self getValidSection:originalIndexPath.section];
-        
+        DTSectionModel * section = [self getValidSection:originalIndexPath.section];
+
         [section.objects replaceObjectAtIndex:originalIndexPath.row
                                    withObject:replacingItem];
     }
-    else {
-        if (self.loggingEnabled) {
-            NSLog(@"DTMemoryStorage: failed to replace item %@ at indexPath: %@",replacingItem,originalIndexPath);
+    else
+    {
+        if (self.loggingEnabled)
+        {
+            NSLog(@"DTMemoryStorage: failed to replace item %@ at indexPath: %@", replacingItem, originalIndexPath);
         }
         return;
     }
-    
+
     [self.currentUpdate.updatedRowIndexPaths addObject:originalIndexPath];
-    
+
     [self finishUpdate];
 }
 
@@ -307,17 +313,19 @@
 - (void)removeItem:(NSObject *)item
 {
     [self startUpdate];
-    
+
     NSIndexPath * indexPath = [self indexPathForItem:item];
-    
+
     if (indexPath)
     {
         DTSectionModel * section = [self getValidSection:indexPath.section];
         [section.objects removeObjectAtIndex:indexPath.row];
     }
-    else {
-        if (self.loggingEnabled) {
-            NSLog(@"DTMemoryStorage: item to delete: %@ was not found",item);
+    else
+    {
+        if (self.loggingEnabled)
+        {
+            NSLog(@"DTMemoryStorage: item to delete: %@ was not found", item);
         }
         return;
     }
@@ -325,22 +333,24 @@
     [self finishUpdate];
 }
 
--(void)removeItemsAtIndexPaths:(NSArray *)indexPaths
+- (void)removeItemsAtIndexPaths:(NSArray *)indexPaths
 {
     [self startUpdate];
     for (NSIndexPath * indexPath in indexPaths)
     {
         id object = [self objectAtIndexPath:indexPath];
-        
+
         if (object)
         {
             DTSectionModel * section = [self getValidSection:indexPath.section];
             [section.objects removeObjectAtIndex:indexPath.row];
             [self.currentUpdate.deletedRowIndexPaths addObject:indexPath];
         }
-        else {
-            if (self.loggingEnabled) {
-                NSLog(@"DTMemoryStorage: item to delete was not found at indexPath : %@ ",indexPath);
+        else
+        {
+            if (self.loggingEnabled)
+            {
+                NSLog(@"DTMemoryStorage: item to delete was not found at indexPath : %@ ", indexPath);
             }
         }
     }
@@ -350,13 +360,13 @@
 - (void)removeItems:(NSArray *)items
 {
     [self startUpdate];
-    
+
     NSArray * indexPaths = [self indexPathArrayForItems:items];
-    
+
     for (NSObject * item in items)
     {
-        NSIndexPath *indexPath = [self indexPathForItem:item];
-        
+        NSIndexPath * indexPath = [self indexPathForItem:item];
+
         if (indexPath)
         {
             DTSectionModel * section = [self getValidSection:indexPath.section];
@@ -369,21 +379,21 @@
 
 #pragma  mark - Sections
 
--(void)deleteSections:(NSIndexSet *)indexSet
+- (void)deleteSections:(NSIndexSet *)indexSet
 {
     [self startUpdate];
     // Update datasource
     [self.sections removeObjectsAtIndexes:indexSet];
-    
+
     // Update interface
     [self.currentUpdate.deletedSectionIndexes addIndexes:indexSet];
-    
+
     [self finishUpdate];
 }
 
 #pragma mark - Search
 
--(NSArray *)itemsInSection:(NSInteger)sectionNumber
+- (NSArray *)itemsInSection:(NSUInteger)sectionNumber
 {
     if ([self.sections count] > sectionNumber)
     {
@@ -396,26 +406,28 @@
     }
 }
 
--(id)itemAtIndexPath:(NSIndexPath *)indexPath
+- (id)itemAtIndexPath:(NSIndexPath *)indexPath
 {
     NSArray * section = nil;
     if (indexPath.section < [self.sections count])
     {
         section = [self itemsInSection:indexPath.section];
     }
-    else {
+    else
+    {
         if (self.loggingEnabled)
         {
             NSLog(@"DTMemoryStorage: Section not found while searching for item");
         }
         return nil;
     }
-    
+
     if (indexPath.row < [section count])
     {
         return [section objectAtIndex:indexPath.row];
     }
-    else {
+    else
+    {
         if (self.loggingEnabled)
         {
             NSLog(@"DTMemoryStorage: Row not found while searching for item");
@@ -424,13 +436,13 @@
     }
 }
 
--(NSIndexPath *)indexPathForItem:(NSObject *)item
+- (NSIndexPath *)indexPathForItem:(NSObject *)item
 {
-    for (NSInteger sectionNumber=0; sectionNumber<self.sections.count; sectionNumber++)
+    for (NSUInteger sectionNumber = 0; sectionNumber < self.sections.count; sectionNumber++)
     {
-        NSArray *rows = [self.sections[sectionNumber] objects];
-        NSInteger index = [rows indexOfObject:item];
-        
+        NSArray * rows = [self.sections[sectionNumber] objects];
+        NSUInteger index = [rows indexOfObject:item];
+
         if (index != NSNotFound)
         {
             return [NSIndexPath indexPathForRow:index inSection:sectionNumber];
@@ -439,29 +451,30 @@
     return nil;
 }
 
--(DTSectionModel *)sectionAtIndex:(NSInteger)sectionNumber
+- (DTSectionModel *)sectionAtIndex:(NSUInteger)sectionNumber
 {
     [self startUpdate];
     DTSectionModel * section = [self getValidSection:sectionNumber];
     [self finishUpdate];
-    
+
     return section;
 }
 
 #pragma mark - private
 
--(DTSectionModel *)getValidSection:(NSUInteger)sectionNumber
+- (DTSectionModel *)getValidSection:(NSUInteger)sectionNumber
 {
     if (sectionNumber < self.sections.count)
     {
         return self.sections[sectionNumber];
     }
-    else {
-        for (NSInteger i = self.sections.count; i <= sectionNumber ; i++)
+    else
+    {
+        for (NSInteger i = self.sections.count; i <= sectionNumber; i++)
         {
             DTSectionModel * section = [DTSectionModel new];
             [self.sections addObject:section];
-            
+
             [self.currentUpdate.insertedSectionIndexes addIndex:i];
         }
         return [self.sections lastObject];
@@ -469,11 +482,11 @@
 }
 
 //This implementation is not optimized, and may behave poorly with lot of sections
--(NSArray *)indexPathArrayForItems:(NSArray *)items
+- (NSArray *)indexPathArrayForItems:(NSArray *)items
 {
     NSMutableArray * indexPaths = [[NSMutableArray alloc] initWithCapacity:[items count]];
-    
-    for (NSInteger i=0; i<[items count]; i++)
+
+    for (NSInteger i = 0; i < [items count]; i++)
     {
         NSIndexPath * foundIndexPath = [self indexPathForItem:[items objectAtIndex:i]];
         if (!foundIndexPath)
@@ -481,10 +494,11 @@
             if (self.loggingEnabled)
             {
                 NSLog(@"DTMemoryStorage: object %@ not found",
-                      [items objectAtIndex:i]);
+                        [items objectAtIndex:i]);
             }
         }
-        else {
+        else
+        {
             [indexPaths addObject:foundIndexPath];
         }
     }
@@ -493,30 +507,30 @@
 
 #pragma mark - helpers
 
--(NSString *)classStringForClass:(Class)class
+- (NSString *)classStringForClass:(Class)class
 {
     NSString * classString = NSStringFromClass(class);
-    
+
     if ([classString isEqualToString:@"__NSCFConstantString"] ||
-        [classString isEqualToString:@"__NSCFString"] ||
-        class == [NSMutableString class])
+            [classString isEqualToString:@"__NSCFString"] ||
+            class == [NSMutableString class])
     {
         return @"NSString";
     }
     if ([classString isEqualToString:@"__NSCFNumber"] ||
-        [classString isEqualToString:@"__NSCFBoolean"])
+            [classString isEqualToString:@"__NSCFBoolean"])
     {
         return @"NSNumber";
     }
     if ([classString isEqualToString:@"__NSDictionaryI"] ||
-        [classString isEqualToString:@"__NSDictionaryM"] ||
-        class == [NSMutableDictionary class])
+            [classString isEqualToString:@"__NSDictionaryM"] ||
+            class == [NSMutableDictionary class])
     {
         return @"NSDictionary";
     }
     if ([classString isEqualToString:@"__NSArrayI"] ||
-        [classString isEqualToString:@"__NSArrayM"] ||
-        class == [NSMutableArray class])
+            [classString isEqualToString:@"__NSArrayM"] ||
+            class == [NSMutableArray class])
     {
         return @"NSArray";
     }
